update customer controller with proper message and added api response class

+++ src/main/java/com/example/BankingApp/Customer/CustomerController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,9 +15,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.BankingApp.Response.ApiResponse;
+
 
 @RestController
-@RequestMapping("/customers")
+@RequestMapping("/api")
 public class CustomerController {
 	
 	@Autowired
@@ -27,35 +30,77 @@ public class CustomerController {
 		return "Welcome to Banking App";
 	}
 	
-	@GetMapping("/all")
-	public List<CustomerEntity> getAllCustomers() {
-		return customerService.getAllCustomers();
+	@GetMapping("/all/customers")
+	public ResponseEntity<?> getAllCustomers() {
+		List<CustomerEntity> customers = customerService.getAllCustomers();
+		if(customers.isEmpty()) {
+			return ResponseEntity
+				.status(HttpStatus.OK)
+				.body(new ApiResponse<>("Customers are not available!!!", customers));
+		}
+		else {
+			return ResponseEntity
+	                .status(HttpStatus.OK)
+	                .body(new ApiResponse<>("Customers fetched Successfully!!!", customers));
+		} 
 	}
 	
 	@PostMapping("/new")
-	public CustomerEntity createCustomer(@RequestBody CustomerEntity customerEntity) {
-		return customerService.saveCustomer(customerEntity);
+	public ResponseEntity<String> createCustomer(@RequestBody CustomerEntity customerEntity) {
+		customerService.saveCustomer(customerEntity);
+		return ResponseEntity.ok("Customer Saved Successfully!!!");
 	}
 	
-	@GetMapping("/{customerId}")
-	public ResponseEntity<CustomerEntity> getCustomerById(@PathVariable Long customerId){
+	@GetMapping("/customer/{customerId}")
+	public ResponseEntity<?> getCustomerById(@PathVariable Long customerId){
 		Optional<CustomerEntity> customerById = customerService.getCustomerById(customerId);
-		return customerById.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+		if(customerById.isPresent()) {
+			return ResponseEntity.ok(customerById.get());
+		}
+		else
+		{
+			return ResponseEntity.
+					status(HttpStatus.NOT_FOUND).
+					body("Customer with Id " + customerId + " not found!!!");		
+		}
 	}
 
-	@PutMapping("/{customerId}")
-	public ResponseEntity<CustomerEntity> updateCustomerById(@PathVariable Long customerId, @RequestBody CustomerEntity customerEntity){
-		try {
-			CustomerEntity updatedPatientById = customerService.updateCustomerById(customerId, customerEntity);
-			return ResponseEntity.ok(updatedPatientById);
-		}catch (Exception e) {
-			return ResponseEntity.notFound().build();
+	@PutMapping("/customer/{customerId}")
+	public ResponseEntity<?> updateCustomerById(@PathVariable Long customerId, @RequestBody CustomerEntity customerEntity){
+		Optional<CustomerEntity> existingCustomer = customerService.findById(customerId);
+		if(existingCustomer.isPresent()) {
+			CustomerEntity customer = existingCustomer.get();
+			customer.setFname(customerEntity.getFname());
+			customer.setLname(customerEntity.getLname());
+			customer.setEmail(customerEntity.getEmail());
+			customer.setPhoneNo(customerEntity.getPhoneNo());
+			customer.setAddress(customerEntity.getAddress());
+			customer.setDob(customerEntity.getDob());
+			customer.setPanNumber(customerEntity.getPanNumber());
+			customer.setStatus(customerEntity.getStatus());
+			customer.setCreatedDate(customerEntity.getCreatedDate());
+			customerService.saveCustomer(customer);
+			return ResponseEntity.ok(new ApiResponse<>("Customer Updated Successfully!!", customer));
+		}
+		else {
+			return ResponseEntity
+					.status(HttpStatus.NOT_FOUND)
+					.body(new ApiResponse<>("Customer not found with Id " +customerId, null));
 		}
 	}
 		
-	@DeleteMapping("/{customerId}")
-	public ResponseEntity<Void> deleteCustomerById(@PathVariable Long customerId){
-		customerService.deleteCustomerById(customerId);
-		return ResponseEntity.noContent().build();
+	@DeleteMapping("/customer/{customerId}")
+	public ResponseEntity<?> deleteCustomerById(@PathVariable Long customerId){
+		Optional<CustomerEntity> existingCustomer = customerService.findById(customerId);
+		if(existingCustomer.isPresent()) {
+			customerService.deleteCustomerById(customerId);
+			return ResponseEntity.ok(new ApiResponse<>("Customer Deleted Successfully!!!!", null));
+		}
+		else {
+			return ResponseEntity
+					.status(HttpStatus.NOT_FOUND)
+					.body(new ApiResponse<>("Customer not found with ID: " + customerId, null));
+		}
 	}
 }
+
